init yui
So we finally realized kiku-style OO is bad.
Took a page from 490; only universally quantified types make any sense

Starting with the traditional singly linked list
and (unfortunately) vector (to use qsort/bsearch for a lookup table).

The idea behind yui:
substitute the type variable during the build process.
modules come with a header file that has basic init/free functions
and prototypes for the to-be-created object files which are only
linked to if used (tracked by build script).
the rest of the .c files in the module directory are functions
that are not built if not used.

lots of work needs to be done, especially with the build script...